services/inference/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
Pillow==11.1.0
exifread==3.0.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
pytest==8.3.4
transformers==4.47.1
torch==2.5.1
